Switch to using single ISO/GPT image
It was not really possible to boot GRUB from EFI El-Torito partitioned
disk image. And it was not really supported by the EFI specification
(the El-Torito image should be 'no emulation' and not partitioned).

On the other hande, xorriso can build ISO images which are also a
GPT-partitioned images. And with a bit of hacking I managed to make
it bootable in four ways:

– legacy boot from CD
– legacy boot from HDD
- efi boot from CD
- efi boot from HDD

The image does not contain the PLD NR system yet, though.