Clean up after initramfs
Most of _init.cpi files must stay as they are not separately packaged in
any of the squashfs images, but _net.cpi is optional, so the files must
be provided in the squashfs modules, the initramfs copies may be safely
removed.

Also, the init scripts from _init.cpi are removed, as they are not
needed in the final system.